Update README.md
This is a Stochastic Unit Commitmend about the electric grid and is specialised to resears the shaddow prices of inertia. There are different thermal Generators and the uncertaintie of wind generation. This programm is based on Matlab2024b version. 

You need YALMIP and also Gurobi for working with this code. You will find some informations in own research about how to install both.


This Stochastic Unit Commitment ist based on the work of:
 
L. Badesa, F. Teng, and G. Strbac, "Simultaneous Scheduling of Multiple Frequency Services in Stochastic Unit Commitment," IEEE Transactions on Power Systems, vol. 34, no. 5, pp. 3858-3868, 2019.